Update game_logic.py
Added a function/logic to hold off on adding new tiles if the board hasn't changed with the last move (i.e. the tiles are stuck in place) akin to traditional 2048 logic.

Not sure if my method is the most time/space efficient, but it seems to work and not affect game performance.